Refactor move getMacroOptions into it's own folder
