+++ lib/getMacroOptions/__mocks__/fs.js
@@ -1,0 +1,12 @@
+/* eslint-env jest */
+
+const fs = jest.genMockFromModule('fs')
+
+let mockedData
+
+fs.__setMockData = data => {
+  mockedData = JSON.stringify(data)
+}
+fs.readFileSync = () => mockedData
+
+module.exports = fs
