+++ lib/metalsmith.js
@@ -29,6 +29,7 @@ const colours = require('../lib/colours.js') // get colours data
 const fileHelper = require('../lib/file-helper.js') // helper function to operate on files
 const paths = require('../config/paths.json') // specify paths to main working directories
 const highlighter = require('./highlighter.js')
+const getMacroOptions = require('./getMacroOptions/index.js')
 
 // store views paths for rendering nunjucks syntax
 const views = [
@@ -213,7 +214,7 @@ module.exports = metalsmith(__dirname) // __dirname defined by node.js: name of 
         getNunjucksCode: fileHelper.getNunjucksCode,
         getHTMLCode: fileHelper.getHTMLCode,
         getFingerprint: fileHelper.getFingerprint,
-        getMacroOptions: fileHelper.getMacroOptions
+        getMacroOptions: getMacroOptions
       },
       filters: {
         highlight: highlighter
